fibonacci.rs: one integer type, not two

+++ scenarios/primitives/loops/fibonacci/src/fibonacci.rs
@@ -1,6 +1,6 @@
-fn fib_iter(n: u32) -> u64 {
+fn fib_iter(n: u64) -> u64 {
     if n < 2 {
-        return n as u64;
+        return n;
     }
     let mut a: u64 = 0;
     let mut b: u64 = 1;
